% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initSpatialExtension.R
\name{initSpatialExtension}
\alias{initSpatialExtension}
\title{Enable spatialite capabilities for an existing sqlite library}
\usage{
initSpatialExtension(con)
}
\arguments{
\item{con}{A RSQLite connection link}
}
\description{
Currently based on theloadable extension from Peter Schmiedeskamp
which has a 'Artistic-2.0' license
}
\author{
Martin Jung
}

