% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initSpatialExtension.R
\name{is_spatial}
\alias{is_spatial}
\title{Quick query if spatiallite is loaded}
\usage{
is_spatial(con)
}
\arguments{
\item{con}{A RSQLite connection link}
}
\description{
Quick query if spatiallite is loaded
}
\author{
Martin Jung
}
\keyword{internal}

